+++ sdk/communication/communication-call-automation/src/callMedia.ts
@@ -23,6 +23,7 @@ import {
   UnholdRequest,
   StartMediaStreamingRequest,
   StopMediaStreamingRequest,
+  SummarizeCallRequest,
 } from "./generated/src/index.js";
 import { KnownPlaySourceType, KnownRecognizeInputType } from "./generated/src/index.js";
 
@@ -51,6 +52,7 @@ import type {
   StopMediaStreamingOptions,
   PlayToAllOptions,
   UpdateTranscriptionOptions,
+  SummarizeCallOptions,
 } from "./models/options.js";
 import type { KeyCredential, TokenCredential } from "@azure/core-auth";
 import type { SendDtmfTonesResult } from "./models/responses.js";
@@ -230,6 +232,8 @@ export class CallMedia {
           : 5,
         targetParticipant: serializeCommunicationIdentifier(targetParticipant),
         speechLanguage: recognizeOptions.speechLanguage,
+        speechLanguages: recognizeOptions.speechLanguages,
+        enableSentimentAnalysis: recognizeOptions.enableSentimentAnalysis,
         speechRecognitionModelEndpointId: recognizeOptions.speechRecognitionModelEndpointId,
         choices: recognizeOptions.choices,
       };
@@ -261,6 +265,8 @@ export class CallMedia {
         targetParticipant: serializeCommunicationIdentifier(targetParticipant),
         speechOptions: speechOptions,
         speechLanguage: recognizeOptions.speechLanguage,
+        speechLanguages: recognizeOptions.speechLanguages,
+        enableSentimentAnalysis: recognizeOptions.enableSentimentAnalysis,
         speechRecognitionModelEndpointId: recognizeOptions.speechRecognitionModelEndpointId,
       };
       return {
@@ -298,6 +304,8 @@ export class CallMedia {
         targetParticipant: serializeCommunicationIdentifier(targetParticipant),
         speechOptions: speechOptions,
         dtmfOptions: dtmfOptionsInternal,
+        speechLanguages: recognizeOptions.speechLanguages,
+        enableSentimentAnalysis: recognizeOptions.enableSentimentAnalysis,
         speechRecognitionModelEndpointId: recognizeOptions.speechRecognitionModelEndpointId,
       };
       return {
@@ -507,6 +515,7 @@ export class CallMedia {
 
   /**
    * Starts transcription in the call
+   * @param locales - List of languages for Language Identification.
    * @param options - Additional attributes for start transcription.
    */
   public async startTranscription(options: StartTranscriptionOptions = {}): Promise<void> {
@@ -515,6 +524,10 @@ export class CallMedia {
       operationContext: options.operationContext,
       speechModelEndpointId: options.speechRecognitionModelEndpointId,
       operationCallbackUri: options.operationCallbackUrl,
+      piiRedactionOptions: options.piiRedactionOptions,
+      locales: options.locales,
+      enableSentimentAnalysis: options.enableSentimentAnalysis,
+      summarizationOptions: options.summarizationOptions,
     };
     return this.callMedia.startTranscription(this.callConnectionId, startTranscriptionRequest, {});
   }
@@ -533,17 +546,19 @@ export class CallMedia {
 
   /**
    * Update transcription language.
-   * @param locale - Defines new locale for transcription.
    */
   public async updateTranscription(
     locale: string,
-    options?: UpdateTranscriptionOptions,
+    options: UpdateTranscriptionOptions = {},
   ): Promise<void> {
     const updateTranscriptionRequest: UpdateTranscriptionRequest = {
       locale: locale,
-      speechModelEndpointId: options?.speechRecognitionModelEndpointId,
-      operationContext: options?.operationContext,
-      operationCallbackUri: options?.operationCallbackUrl,
+      speechModelEndpointId: options.speechRecognitionModelEndpointId,
+      operationContext: options.operationContext,
+      operationCallbackUri: options.operationCallbackUrl,
+      piiRedactionOptions: options.piiRedactionOptions,
+      enableSentimentAnalysis: options.enableSentimentAnalysis,
+      summarizationOptions: options.summarizationOptions,
     };
     return this.callMedia.updateTranscription(
       this.callConnectionId,
@@ -551,6 +566,23 @@ export class CallMedia {
       {},
     );
   }
+
+  /**
+   * Summarize call details.
+   * @param options - Additional attributes for summarize call.
+   */
+  public async summarizeCall(options: SummarizeCallOptions = {}): Promise<void> {
+    const summarizeCallRequest: SummarizeCallRequest = {
+      operationContext: options.operationContext,
+      operationCallbackUri: options.operationCallbackUrl,
+      summarizationOptions: {
+        enableEndCallSummary: options.summarizationOptions?.enableEndCallSummary,
+        locale: options.summarizationOptions?.locale,
+      },
+    };
+    return this.callMedia.summarizeCall(this.callConnectionId, summarizeCallRequest, {});
+  }
+
   /**
    * Starts media streaming in the call.
    * @param options - Additional attributes for start media streaming.