+++ sdk/communication/communication-call-automation/src/models/models.ts
@@ -16,6 +16,7 @@ import type {
   MediaStreamingAudioChannelType,
   MediaStreamingContentType,
   AudioFormat,
+  RedactionType,
 } from "../generated/src/index.js";
 
 export {
@@ -255,7 +256,12 @@ export interface SipCustomHeader extends CustomCallingContextHeader {
 export type SipHeaderPrefix = "X-" | "X-MS-Custom-";
 
 /** Custom Calling Context */
-export type CustomCallingContext = (VoipHeader | SipUserToUserHeader | SipCustomHeader)[];
+export type CustomCallingContext = (
+  | VoipHeader
+  | SipUserToUserHeader
+  | SipCustomHeader
+  | TeamsPhoneCallDetails
+)[];
 
 /** AI options for the call. */
 export interface CallIntelligenceOptions {
@@ -285,7 +291,7 @@ export interface MediaStreamingOptions {
 /** Options for media streaming. */
 export interface TranscriptionOptions {
   /** Specifies the Locale used for transcription, e.g., en-CA or en-AU. */
-  locale: string;
+  locale?: string;
   /** Polymorphic discriminator, which specifies the different types this object can be */
   transportType: "websocket";
   /** The URL used for live transcription transport. */
@@ -296,4 +302,84 @@ export interface TranscriptionOptions {
   startTranscription?: boolean;
   /** Enables intermediate results for the transcribed speech. */
   enableIntermediateResults?: boolean;
+  /** PII redaction configuration options. */
+  piiRedactionOptions?: PiiRedactionOptions;
+  /** Indicating if sentiment analysis should be enabled. */
+  enableSentimentAnalysis?: boolean;
+  /**
+   * List of locales for Language Identification.
+   * Supports upto 4 locales in the format: ["en-us", "fr-fr", "hi-in"] etc.
+   */
+  locales?: string[];
+  /** Summarization configuration options. */
+  summarizationOptions?: SummarizationOptions;
+}
+
+/** PII redaction configuration options. */
+export interface PiiRedactionOptions {
+  /** Value indicating if PII redaction is enabled. */
+  enable?: boolean;
+  /** Value indicating the PII redaction type to be used. */
+  redactionType?: RedactionType;
+}
+
+/** Configuration options for call summarization. */
+export interface SummarizationOptions {
+  /** Indicating whether end call summary should be enabled. */
+  enableEndCallSummary?: boolean;
+  /** Locale for summarization (e.g., en-US). */
+  locale?: string;
+}
+
+/** Teams phone call details */
+export interface TeamsPhoneCallDetails {
+  kind: "teamsPhoneCallDetails";
+  /** Container for details relating to the original caller of the call */
+  teamsPhoneCallerDetails?: TeamsPhoneCallerDetails;
+  /** Container for details relating to the entity responsible for the creation of these call details */
+  teamsPhoneSourceDetails?: TeamsPhoneSourceDetails;
+  /** Id to exclusively identify this call session. IVR will use this for their telemetry/reporting. */
+  sessionId?: string;
+  /** The intent of the call */
+  intent?: string;
+  /** A very short description (max 48 chars) of the reason for the call. To be displayed in Teams CallNotification */
+  callTopic?: string;
+  /** A summary of the call thus far. It will be displayed on a side panel in the Teams UI */
+  callContext?: string;
+  /** Url for fetching the transcript of the call */
+  transcriptUrl?: string;
+  /** Sentiment of the call thus far */
+  callSentiment?: string;
+  /** Recommendations for resolving the issue based on the customer's intent and interaction history */
+  suggestedActions?: string;
+}
+
+/** Container for details relating to the original caller of the call */
+export interface TeamsPhoneCallerDetails {
+  /** Caller's ID */
+  caller: CommunicationIdentifier;
+  /** Caller's name */
+  name: string;
+  /** Caller's phone number */
+  phoneNumber: string;
+  /** Caller's record ID (ex in CRM) */
+  recordId?: string;
+  /** Caller's screen pop URL */
+  screenPopUrl?: string;
+  /** Flag indicating whether the caller was authenticated */
+  isAuthenticated?: boolean;
+  /** A set of key value pairs (max 10, any additional entries would be ignored) which a bot author wants to pass to the Teams Client for display to the agent */
+  additionalCallerInformation?:  Record<string, string>;
+}
+
+/** Container for details relating to the entity responsible for the creation of these call details */
+export interface TeamsPhoneSourceDetails {
+  /** ID of the source entity passing along the call details (ex. Application Instance ID of - CQ/AA) */
+  source: CommunicationIdentifier;
+  /** Language of the source entity passing along the call details, passed in the ISO-639 standard */
+  language: string;
+  /** Status of the source entity passing along the call details */
+  status: string;
+  /** Intended targets of the source entity passing along the call details */
+  intendedTargets?: Record<string, CommunicationIdentifier>;
 }