+++ sdk/communication/communication-call-automation/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Release History
 
-## 1.5.0-beta.3 (Unreleased)
+## 1.6.0-beta.1 (Unreleased)
+
+### Features Added
+- Added support for move participant in a call to another call.
+- Added support for Get recording result.
+- Added support for Teams phone call details with custom calling context.
+- Added support for PiiRedaction and summarization in transcription.
+- Added support for sentiments analysis in speech and choice recognition.
+
+## 1.5.0 (2025-07-25)
 
 ### Features Added